+++ src/c_cmds.c
@@ -413,6 +413,7 @@ static dboolean turbo_cvar_func1(char *, char *, char *, char *);
 static void turbo_cvar_func2(char *, char *, char *, char *);
 static dboolean units_cvar_func1(char *, char *, char *, char *);
 static void units_cvar_func2(char *, char *, char *, char *);
+static dboolean vid_capfps_cvar_func1(char *, char *, char *, char *);
 static void vid_capfps_cvar_func2(char *, char *, char *, char *);
 static void vid_display_cvar_func2(char *, char *, char *, char *);
 static void vid_fullscreen_cvar_func2(char *, char *, char *, char *);
@@ -756,8 +757,8 @@ consolecmd_t consolecmds[] =
         "The units used in the <b>playerstats</b> CCMD (<b>imperial</b> or <b>metric</b>)."),
     CVAR_STR(version, "", null_func1, str_cvars_func2, CF_READONLY,
         "<i><b>"PACKAGE_NAME"'s</b></i> version."),
-    CVAR_INT(vid_capfps, "", int_cvars_func1, vid_capfps_cvar_func2, CF_NONE, NOALIAS,
-        "The number of frames per second at which to cap the\nframerate (<b>35</b> to <b>1,000</b>)."),
+    CVAR_INT(vid_capfps, "", vid_capfps_cvar_func1, vid_capfps_cvar_func2, CF_NONE, NOALIAS,
+        "The frames per second at which to cap the framerate (<b>off</b>, or\n<b>35</b> to <b>1,000</b>)."),
      CVAR_INT(vid_display, "", int_cvars_func1, vid_display_cvar_func2, CF_NONE, NOALIAS,
         "The display used to render the game."),
 #if !defined(WIN32)
@@ -4322,13 +4323,37 @@ static void units_cvar_func2(char *cmd, char *parm1, char *parm2, char *parm3)
 //
 // vid_capfps CVAR
 //
+static dboolean vid_capfps_cvar_func1(char *cmd, char *parm1, char *parm2, char *parm3)
+{
+    int value = -1;
+
+    if (!*parm1 || M_StringCompare(parm1, "off"))
+        return true;
+
+    sscanf(parm1, "%10i", &value);
+
+    return (value >= vid_capfps_min && value <= vid_capfps_max);
+}
+
 static void vid_capfps_cvar_func2(char *cmd, char *parm1, char *parm2, char *parm3)
 {
     int vid_capfps_old = vid_capfps;
 
-    int_cvars_func2(cmd, parm1, "", "");
-    if (vid_capfps != vid_capfps_old)
-        I_RestartGraphics();
+    if (M_StringCompare(parm1, "off"))
+    {
+        vid_capfps = 0;
+        if (vid_capfps != vid_capfps_old)
+        {
+            M_SaveCVARs();
+            I_RestartGraphics();
+        }
+    }
+    else
+    {
+        int_cvars_func2(cmd, parm1, "", "");
+        if (vid_capfps != vid_capfps_old)
+            I_RestartGraphics();
+    }
 }
 
 //