+++ src/r_things.c
@@ -698,7 +698,7 @@ void R_ProjectSprite(mobj_t *thing)
     fixed_t             topoffset;
 
     // [AM] Interpolate between current and last position, if prudent.
-    if (vid_capfps > TICRATE
+    if (vid_capfps != TICRATE
         // Don't interpolate if the mobj did something
         // that would necessitate turning it off for a tic.
         && thing->interp
@@ -1243,7 +1243,7 @@ static void R_DrawPSprite(pspdef_t *psp, dboolean invisibility)
     }
 
     // e6y: interpolation for weapon bobbing
-    if (vid_capfps > TICRATE)
+    if (vid_capfps != TICRATE)
     {
         typedef struct
         {